+++ Application/Wrappers/Response.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Application.Wrappers
+{
+    public class Response<T>
+    {
+        public Response()
+        {
+        }
+        public Response(T data)
+        {
+            Succeeded = true;
+            Message = string.Empty;
+            Errors = null;
+            Data = data;
+        }
+        public T Data { get; set; }
+        public bool Succeeded { get; set; }
+
+        public string[] Errors { get; set; }
+        public string Message { get; set; }
+    }
+}